pyproject.toml
[project]
name = "fastapi-sqlmodel-demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "aiosqlite>=0.22.1",
    "apiexception>=0.2.1",
    "asyncpg>=0.31.0",
    "fastapi[standard]>=0.128.0",
    "loguru>=0.7.3",
    "pwdlib>=0.3.0",
    "pyjwt>=2.10.1",
    "sqlalchemy[asyncio]>=2.0.45",
    "sqlmodel==0.0.31",
]


# 可选依赖分组 - 专门存放测试/开发依赖（生产环境不会主动安装）
[project.optional-dependencies]
# 命名为 test，可自定义（如 dev、test-utils）
test = [
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
]
